fix(guard): harden activity privacy endpoints
