+++ tests/test_guard_command_activity_deletion_privacy.py
@@ -15,12 +15,10 @@ from typing import cast
 
 import pytest
 
-from codex_plugin_scanner.guard.approval_gate import update_settings as update_approval_gate_settings
 from codex_plugin_scanner.guard.cli.commands_support_command_activity import (
     record_pre_hook_command_activity_best_effort,
 )
 from codex_plugin_scanner.guard.daemon.server import GuardDaemonServer
-from codex_plugin_scanner.guard.local_dashboard_session import build_local_dashboard_session_token
 from codex_plugin_scanner.guard.runtime.command_activity_api_contract import (
     COMMAND_ACTIVITY_API_SCHEMA_VERSION,
     CommandActivityFeedbackLabel,
@@ -38,10 +36,10 @@ from codex_plugin_scanner.guard.store_command_activity_maintenance_schema import
 )
 from codex_plugin_scanner.guard.store_command_activity_privacy import (
     COMMAND_ACTIVITY_DIAGNOSTICS_SCHEMA_VERSION,
+    _stable_distinct,
 )
-from tests.guard_command_activity_api_support import json_request, seed
+from tests.guard_command_activity_api_support import seed
 
-_PASSWORD = "correct horse battery staple"
 _COMMAND_TABLES = (
     "command_activity",
     "command_activity_matches",
@@ -58,25 +56,6 @@ _COMMAND_TABLES = (
 
 def _store(tmp_path: Path) -> GuardStore:
     return GuardStore(tmp_path / "guard-home", prime_policy_integrity=False)
-
-
-def _enable_gate(store: GuardStore) -> None:
-    update_approval_gate_settings(
-        store.guard_home,
-        {
-            "enabled": True,
-            "new_password": _PASSWORD,
-            "confirm_password": _PASSWORD,
-            "cooldown_seconds": 0,
-        },
-    )
-
-
-def _dashboard_token(daemon: GuardDaemonServer) -> str:
-    return build_local_dashboard_session_token(
-        auth_token=daemon._server.auth_token,
-        surface="dashboard",
-    )
 
 
 def _all_command_rows(store: GuardStore) -> dict[str, list[list[object]]]:
@@ -207,6 +186,9 @@ def test_diagnostics_has_an_exact_bounded_allowlist(tmp_path: Path) -> None:
     for forbidden in ("activity:01", "receipt:", "occurred_at", "last_error_at", "digest"):
         assert forbidden not in serialized
 
+    with store._connect() as connection:
+        assert _stable_distinct(connection, "command_activity", "harness", allowed=frozenset()) == []
+
 
 def test_diagnostics_omits_corrupt_unbounded_identifiers(tmp_path: Path) -> None:
     store = _store(tmp_path)
@@ -266,35 +248,6 @@ def test_diagnostics_omits_corrupt_unbounded_identifiers(tmp_path: Path) -> None
     assert missing_health["error_classes"] == []
 
 
-def test_diagnostics_export_requires_dashboard_auth_and_hosted_allowlist(tmp_path: Path) -> None:
-    store = _store(tmp_path)
-    seed(store)
-    daemon = GuardDaemonServer(store, host="127.0.0.1", port=0)
-    daemon.start()
-    try:
-        unauthenticated = json_request(daemon, "/v1/command-activity/diagnostics")
-        allowed = json_request(
-            daemon,
-            "/v1/command-activity/diagnostics",
-            token=_dashboard_token(daemon),
-            origin="https://hol.org",
-        )
-        forbidden = json_request(
-            daemon,
-            "/v1/command-activity/diagnostics",
-            token=_dashboard_token(daemon),
-            origin="https://attacker.example",
-        )
-    finally:
-        daemon.stop()
-
-    assert unauthenticated[0] == 401
-    assert allowed[0] == 200
-    assert allowed[1]["schema_version"] == COMMAND_ACTIVITY_DIAGNOSTICS_SCHEMA_VERSION
-    assert allowed[2]["Access-Control-Allow-Origin"] == "https://hol.org"
-    assert forbidden[:2] == (403, {"error": "forbidden_origin"})
-
-
 def test_clear_is_atomic_and_preserves_unrelated_state(tmp_path: Path) -> None:
     store = _store(tmp_path)
     seed(store)
@@ -349,60 +302,6 @@ def test_clear_is_atomic_and_preserves_unrelated_state(tmp_path: Path) -> None:
             "occurred_at": "2026-07-18T22:00:00+00:00",
         }
     ]
-
-
-def test_delete_route_requires_confirmation_and_high_risk_gate(tmp_path: Path) -> None:
-    store = _store(tmp_path)
-    seed(store)
-    _enable_gate(store)
-    daemon = GuardDaemonServer(store, host="127.0.0.1", port=0)
-    daemon.start()
-    token = _dashboard_token(daemon)
-    try:
-        missing_confirmation = json_request(
-            daemon,
-            "/v1/command-activity",
-            method="DELETE",
-            token=token,
-            payload={},
-        )
-        missing_gate = json_request(
-            daemon,
-            "/v1/command-activity",
-            method="DELETE",
-            token=token,
-            payload={"confirm": "clear-command-activity"},
-        )
-        wrong_gate = json_request(
-            daemon,
-            "/v1/command-activity",
-            method="DELETE",
-            token=token,
-            payload={"confirm": "clear-command-activity", "approval_password": "incorrect password"},
-        )
-        assert store.count_command_activities() == 3
-        cleared = json_request(
-            daemon,
-            "/v1/command-activity",
-            method="DELETE",
-            token=token,
-            payload={"confirm": "clear-command-activity", "approval_password": _PASSWORD},
-        )
-    finally:
-        daemon.stop()
-
-    assert missing_confirmation[:2] == (
-        400,
-        {"error": "confirmation_required", "confirm": "clear-command-activity"},
-    )
-    assert missing_gate[0] == 403
-    assert missing_gate[1]["error"] == "approval_gate_required"
-    assert wrong_gate[0] == 403
-    assert wrong_gate[1]["error"] == "approval_gate_invalid_password"
-    assert cleared[0] == 200
-    deleted = cast(dict[str, object], cleared[1]["deleted"])
-    assert deleted["activities"] == 3
-    assert store.count_command_activities() == 0
 
 
 def test_forbidden_sentinels_never_reach_command_activity_surfaces(
